feat: group the table properties by what owns them
