+++ frontend/src/apps/slides/components/TableSection.vue
@@ -32,19 +32,37 @@
 				</template>
 			</Select>
 		</PropertyRow>
+		<PropertyRow
+			label="Banded Rows"
+			class="cursor-pointer"
+			@click="toggleFromRow($event, () => setBandedRows(!activeElement.bandedRows))"
+		>
+			<Switch :modelValue="activeElement.bandedRows || false" @update:modelValue="setBandedRows" />
+		</PropertyRow>
+		<PropertyRow v-if="activeElement.bandedRows" label="Band Color">
+			<ColorPicker
+				:modelValue="activeElement.bandColor || getDefaultBandColor(activeElement.color)"
+				@update:modelValue="bandColor.set"
+				@colordown="bandColor.begin"
+				@colorup="bandColor.commit"
+			/>
+		</PropertyRow>
 	</Section>
 </template>
 
 <script setup>
 import { computed } from 'vue'
 
-import { Select } from 'frappe-ui'
+import { Select, Switch } from 'frappe-ui'
 
+import ColorPicker from '@/apps/slides/components/controls/ColorPicker.vue'
 import NumberControl from '@/apps/slides/components/controls/NumberControl.vue'
 import PropertyRow from '@/apps/slides/components/controls/PropertyRow.vue'
 import Section from '@/apps/slides/components/controls/Section.vue'
 
 import { activeElement } from '@/apps/slides/stores/element'
+import { setElementProperty, useElementProperty } from '@/apps/slides/composables/editProperty'
+import { getDefaultBandColor } from '@/apps/slides/utils/color'
 import { chevronClasses } from '@/apps/slides/utils/constants'
 import { getTableSize, getMinTableSize, getTableHeaders } from '@/apps/slides/utils/tableWidths'
 import {
@@ -81,6 +99,15 @@ const setHeaderMode = (value) => {
 	if ((value === 'row' || value === 'both') !== row) toggleHeaderRow()
 	if ((value === 'column' || value === 'both') !== column) toggleHeaderColumn()
 }
+
+// the switch handles its own clicks; the rest of the row forwards to it
+const toggleFromRow = (e, toggle) => {
+	if (!e.target.closest('button')) toggle()
+}
+
+const setBandedRows = (value) => setElementProperty('bandedRows', value)
+
+const bandColor = useElementProperty('bandColor')
 
 const valueClasses = 'block font-text text-base text-ink-gray-8'
 </script>
